requirements/base.txt
asgiref==3.11.0
Django==6.0.1
sqlparse==0.5.5
psycopg2-binary==2.9.10
django-environ==0.11.2
djangorestframework==3.15.2
drf-spectacular==0.28.0
django-auditlog==3.0.0
django-modeltranslation==0.19.10
django-rest-knox==5.0.2
celery==5.4.0
redis==5.2.1
flower==2.0.1
django-import-export==4.3.7
django-import-export-celery==1.7.1
